Ensure adjacent ranges are merged in RegionGroup
